requirements.txt
requests
bs4
python-docx